+++ demos/58-shape-reflecting-rays/index.js
@@ -127,7 +127,7 @@
           1
         );
 
-        if( shape instanceof BezierPath ) {
+        if (shape instanceof BezierPath) {
           console.log("is bezier path");
           var bbox = findBB(shape.bezierCurves[0]);
           draw.rect(
@@ -329,72 +329,89 @@
       return rays;
     };
 
+    function findBB(bCurve) {
+      var P = [
+        { X: bCurve.startPoint.x, Y: bCurve.startPoint.y },
+        { X: bCurve.startControlPoint.x, Y: bCurve.startControlPoint.y },
+        { X: bCurve.endControlPoint.x, Y: bCurve.endControlPoint.y },
+        { X: bCurve.endPoint.x, Y: bCurve.endPoint.y }
+      ];
+      var a, b, c, disc;
+      // var a = 3 * P[3].X - 9 * P[2].X + 9 * P[1].X - 3 * P[0].X;
+      // var b = 6 * P[0].X - 12 * P[1].X + 6 * P[2].X;
+      // var c = 3 * P[1].X - 3 * P[0].X;
+      // var polyX = cubicPoly(
+      //   P[0].X, // bezierCurve.startPoint.x,
+      //   P[1].X, // bezierCurve.startControlPoint.x,
+      //   P[2].X, // bezierCurve.endControlPoint.x,
+      //   P[3].X // bezierCurve.endPoint.x
+      // );
+      // var a = polyX.a;
+      // var b = polyX.b;
+      // var c = polyX.c;
+      // //alert("a "+a+" "+b+" "+c);
+      // // var disc = b * b - 4 * a * c;
+      // // var disc = polyX.b * polyX.b - 4 * polyX.a * polyX.c;
+      // var disc = polyX.b * polyX.b - 4 * polyX.a * polyX.c;
+      // var xl = Math.min(bCurve.endPoint.x, bCurve.startPoint.x); // P[0].X;
+      // var xh = Math.max(bCurve.endPoint.x, bCurve.startPoint.x); // P[0].X;
+      // // if (P[3].X < xl) xl = P[3].X;
+      // // if (P[3].X > xh) xh = P[3].X;
+      // if (disc >= 0) {
+      //   var t1 = (-b + Math.sqrt(disc)) / (2 * a);
+      //   // alert("t1 " + t1);
+      //   if (t1 > 0 && t1 < 1) {
+      //     // var x1 = evalBez(PX, t1);
+      //     var x1 = bCurve.getPointAt(t1).x;
+      //     if (x1 < xl) xl = x1;
+      //     if (x1 > xh) xh = x1;
+      //   }
 
-    function findBB(bezierCurve) {
-      var P = [ 
-        { X : bezierCurve.startPoint.x, Y: bezierCurve.startPoint.y},
-        { X : bezierCurve.startControlPoint.x, Y: bezierCurve.startControlPoint.y},
-        { X : bezierCurve.endControlPoint.x, Y: bezierCurve.endControlPoint.y},
-        { X : bezierCurve.endPoint.x, Y: bezierCurve.endPoint.y}
-            ];
-      var a = 3 * P[3].X - 9 * P[2].X + 9 * P[1].X - 3 * P[0].X;
-      var b = 6 * P[0].X - 12 * P[1].X + 6 * P[2].X;
-      var c = 3 * P[1].X - 3 * P[0].X;
-      //alert("a "+a+" "+b+" "+c);
-      var disc = b * b - 4 * a * c;
-      var xl = P[0].X;
-      var xh = P[0].X;
-      if (P[3].X < xl) xl = P[3].X;
-      if (P[3].X > xh) xh = P[3].X;
-      if (disc >= 0) {
-          var t1 = (-b + Math.sqrt(disc)) / (2 * a);
-          // alert("t1 " + t1);
-          if (t1 > 0 && t1 < 1) {
-              // var x1 = evalBez(PX, t1);
-              var x1 = bezierCurve.getPointAt(t1).x; 
-              if (x1 < xl) xl = x1;
-              if (x1 > xh) xh = x1;
-          }
-  
-          var t2 = (-b - Math.sqrt(disc)) / (2 * a);
-          // alert("t2 " + t2);
-          if (t2 > 0 && t2 < 1) {
-              // var x2 = evalBez(PX, t2);
-              var x2 = bezierCurve.getPointAt(t2).x; 
-              if (x2 < xl) xl = x2;
-              if (x2 > xh) xh = x2;
-          }
-      }
-  
-      a = 3 * P[3].Y - 9 * P[2].Y + 9 * P[1].Y - 3 * P[0].Y;
-      b = 6 * P[0].Y - 12 * P[1].Y + 6 * P[2].Y;
-      c = 3 * P[1].Y - 3 * P[0].Y;
-      disc = b * b - 4 * a * c;
-      var yl = P[0].Y;
-      var yh = P[0].Y;
-      if (P[3].Y < yl) yl = P[3].Y;
-      if (P[3].Y > yh) yh = P[3].Y;
-      if (disc >= 0) {
-          var t1 = (-b + Math.sqrt(disc)) / (2 * a);
-          // alert("t3 " + t1);
-  
-          if (t1 > 0 && t1 < 1) {
-              // var y1 = evalBez(PY, t1);
-              var y1 = bezierCurve.getPointAt(t1).y; // evalBez(PY, t1);
-              if (y1 < yl) yl = y1;
-              if (y1 > yh) yh = y1;
-          }
-  
-          var t2 = (-b - Math.sqrt(disc)) / (2 * a);
-          // alert("t4 " + t2);
-  
-          if (t2 > 0 && t2 < 1) {
-              // var y2 = evalBez(PY, t2);
-              var y2 = bezierCurve.getPointAt(t2).y; 
-              if (y2 < yl) yl = y2;
-              if (y2 > yh) yh = y2;
-          }
-      }
+      //   var t2 = (-b - Math.sqrt(disc)) / (2 * a);
+      //   // alert("t2 " + t2);
+      //   if (t2 > 0 && t2 < 1) {
+      //     // var x2 = evalBez(PX, t2);
+      //     var x2 = bCurve.getPointAt(t2).x;
+      //     if (x2 < xl) xl = x2;
+      //     if (x2 > xh) xh = x2;
+      //   }
+      // }
+      var xMinMax = cubicPolyMinMax(bCurve.startPoint.x, bCurve.startControlPoint.x, bCurve.endControlPoint.x, bCurve.endPoint.x);
+      var xl = xMinMax.min;
+      var xh = xMinMax.max;
+
+      // a = 3 * P[3].Y - 9 * P[2].Y + 9 * P[1].Y - 3 * P[0].Y;
+      // b = 6 * P[0].Y - 12 * P[1].Y + 6 * P[2].Y;
+      // c = 3 * P[1].Y - 3 * P[0].Y;
+      // disc = b * b - 4 * a * c;
+      // var yl = P[0].Y;
+      // var yh = P[0].Y;
+      // if (P[3].Y < yl) yl = P[3].Y;
+      // if (P[3].Y > yh) yh = P[3].Y;
+      // if (disc >= 0) {
+      //   var t1 = (-b + Math.sqrt(disc)) / (2 * a);
+      //   // alert("t3 " + t1);
+
+      //   if (t1 > 0 && t1 < 1) {
+      //     // var y1 = evalBez(PY, t1);
+      //     var y1 = bCurve.getPointAt(t1).y; // evalBez(PY, t1);
+      //     if (y1 < yl) yl = y1;
+      //     if (y1 > yh) yh = y1;
+      //   }
+
+      //   var t2 = (-b - Math.sqrt(disc)) / (2 * a);
+      //   // alert("t4 " + t2);
+
+      //   if (t2 > 0 && t2 < 1) {
+      //     // var y2 = evalBez(PY, t2);
+      //     var y2 = bCurve.getPointAt(t2).y;
+      //     if (y2 < yl) yl = y2;
+      //     if (y2 > yh) yh = y2;
+      //   }
+      // }
+      var yMinMax = cubicPolyMinMax(bCurve.startPoint.y, bCurve.startControlPoint.y, bCurve.endControlPoint.y, bCurve.endPoint.y);
+      var yl = yMinMax.min;
+      var yh = yMinMax.max;
 
       // ctx.moveTo(xl, yl);
       // ctx.lineTo(xl, yh);
@@ -402,8 +419,69 @@
       // ctx.lineTo(xh, yl);
       // ctx.lineTo(xl, yl);
 
-      return Bounds.computeFromVertices( [ { x : xl, y: xl}, { x: xl, y : yh }, { x : xh, y : yh }, { x : xh, y: yl } ] );
+      return Bounds.computeFromVertices([
+        { x: xl, y: yl },
+        { x: xl, y: yh },
+        { x: xh, y: yh },
+        { x: xh, y: yl }
+      ]);
     }
+
+    var cubicPolyMinMax = function (p0, p1, p2, p3) {
+      var polyX = cubicPoly(
+        p0, // P[0].X, // bezierCurve.startPoint.x,
+        p1, // P[1].X, // bezierCurve.startControlPoint.x,
+        p2, // P[2].X, // bezierCurve.endControlPoint.x,
+        p3 // P[3].X // bezierCurve.endPoint.x
+      );
+      var a = polyX.a;
+      var b = polyX.b;
+      var c = polyX.c;
+      //alert("a "+a+" "+b+" "+c);
+      // var disc = b * b - 4 * a * c;
+      // var disc = polyX.b * polyX.b - 4 * polyX.a * polyX.c;
+      var disc = polyX.b * polyX.b - 4 * polyX.a * polyX.c;
+      // var xl = Math.min(bCurve.endPoint.x, bCurve.startPoint.x); // P[0].X;
+      // var xh = Math.max(bCurve.endPoint.x, bCurve.startPoint.x); // P[0].X;
+      var xl = Math.min(p3, p0); // P[0].X;
+      var xh = Math.max(p3, p0); // P[0].X;
+      // if (P[3].X < xl) xl = P[3].X;
+      // if (P[3].X > xh) xh = P[3].X;
+      if (disc >= 0) {
+        var t1 = (-b + Math.sqrt(disc)) / (2 * a);
+        // alert("t1 " + t1);
+        if (t1 > 0 && t1 < 1) {
+          // var x1 = evalBez(PX, t1);
+          // var x1 = bCurve.getPointAt(t1).x;
+          var x1 = CubicBezierCurve.utils.evaluateT(p0, p1, p2, p3, t1); // bCurve.getPointAt(t1).x;
+
+          if (x1 < xl) xl = x1;
+          if (x1 > xh) xh = x1;
+        }
+
+        var t2 = (-b - Math.sqrt(disc)) / (2 * a);
+        // alert("t2 " + t2);
+        if (t2 > 0 && t2 < 1) {
+          // var x2 = evalBez(PX, t2);
+          // var x2 = bCurve.getPointAt(t2).x;
+          var x2 = CubicBezierCurve.utils.evaluateT(p0, p1, p2, p3, t2); //
+          if (x2 < xl) xl = x2;
+          if (x2 > xh) xh = x2;
+        }
+      }
+      return { min: xl, max: xh };
+    };
+
+    var cubicPoly = function (p0, p1, p2, p3) {
+      // var a = 3 * P[3].X - 9 * P[2].X + 9 * P[1].X - 3 * P[0].X;
+      // var b = 6 * P[0].X - 12 * P[1].X + 6 * P[2].X;
+      // var c = 3 * P[1].X - 3 * P[0].X;
+      return {
+        a: 3 * p3 - 9 * p2 + 9 * p1 - 3 * p0,
+        b: 6 * p0 - 12 * p1 + 6 * p2,
+        c: 3 * p1 - 3 * p0
+      };
+    };
 
     // +---------------------------------------------------------------------------------
     // | Render next animation step.